Moving Resource to it's own module.
